refactor(jazz): isolate wire transport adaptation
